fix: long country id to integer

+++ src/main/java/com/pinapp/clients/persistence/crud/CountryCrud.java
@@ -3,5 +3,5 @@ package com.pinapp.clients.persistence.crud;
 import com.pinapp.clients.persistence.entity.CountryEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CountryCrud extends CrudRepository<CountryEntity, Long> {
+public interface CountryCrud extends CrudRepository<CountryEntity, Integer> {
 }
